==> 1_load_openweather_data.ipynb <==
import requests
import datetime
from datetime import datetime, timedelta
import json
import ast
import time


api_key = 'a2b11c4d97b642bc45a2776615179cd5'
#  29.18483, 71.49811 (29°11'05.4"N 71°29'53.2"E) 
lat = 29.18483
lon = 71.49811


def string_to_json_file(dict_string, filename):
    try:
        # Convert string to dictionary using ast.literal_eval
        dictionary = ast.literal_eval(dict_string)
        
        # Write dictionary to JSON file
        with open(filename, 'a') as json_file:
            json.dump(dictionary, json_file, indent=4)
            json_file.write(',\n')
        return True
    except Exception as e:
        print(f"Error: {e}")
        return False



def generate_date_list(start_date, end_date):
    start = datetime.strptime(start_date, '%Y-%m-%d')
    end = datetime.strptime(end_date, '%Y-%m-%d')
    
    date_list = []
    current_date = start
    
    while current_date <= end:
        date_list.append(current_date.strftime('%Y-%m-%d'))
        current_date += timedelta(days=1)
        
    return date_list


# Generate the full date list
dates = generate_date_list('2019-01-01', '2024-12-16')

# Iterate through dates with a limit of 995
stop_index = min(995, len(dates))
last_date = None
filename = "weather_data.json"


for i in range(stop_index):
    
    url = f'https://api.openweathermap.org/data/3.0/onecall/day_summary?lat={lat}&lon={lon}&date={dates[i]}&appid={api_key}'
    get_response = requests.get(url)
    time.sleep(1)
    
    if string_to_json_file(get_response.text, filename):
        print(f"Run #{i+1}, {dates[i]} successfully written to {filename} file.")
    else:
        print("Failed to write JSON file.")


    last_date = dates[i]

# Write the last processed date to a file
if last_date:
    with open('last_processed_date.txt', 'w') as f:
        f.write(f"Stopped at date: {last_date}\n")
        f.write(f"Iteration count: {stop_index}")
    print(f"\nScript stopped at date: {last_date}")
    print(f"Last date has been written to 'last_processed_date.txt'")
