==> 2_read_transform_weather_data.ipynb <==
import pandas as pd
import json

def json_to_dataframe(json_file_path):
    try:
        # Read JSON file
        with open(json_file_path, 'r') as file:
            data = json.load(file)
        
        # Convert to pandas DataFrame
        df = pd.DataFrame(data)
        
        return df
    
    except FileNotFoundError:
        print(f"Error: File '{json_file_path}' not found")
        return None
    except json.JSONDecodeError:
        print("Error: Invalid JSON format")
        return None
    except Exception as e:
        print(f"Error: {str(e)}")
        return None


# Example usage
file_path = 'weather_data.json'
df = json_to_dataframe(file_path)

df.info()

df_start=json_to_dataframe('weather_data_test.json')

final_df = pd.concat([df_start, df], axis=0, ignore_index=True)

final_df.head()

final_df.units.value_counts()

final_df.to_csv('openweathermap_data.csv', index=False)

final_df.temperature[0]
